implementando openfeign no product-service

+++ user-service/src/main/java/br/com/rafaelblomer/business/dtos/ResponseUsuarioDTO.java
@@ -7,5 +7,5 @@ public record ResponseUsuarioDTO(Long idUsuario,
                                  String email,
                                  String telefone,
                                  String cnpj,
-                                 List<Integer> idsEstoques) {
+                                 List<Long> idsEstoques) {
 }
